boards: arm: mimxrt1060_evk: added can fd support
The flexcan3 node is updated to flexcan-fd due to the introduction
of the new flexcan-fd compatible. Previously used flexcan2 node is
now disabled and the flexcan3 node replaces it as the chosen node.
Pins formerly used by the flexcan2 node is now used by flexcan3.
Documentation updated regarding flexcan-fd support.

Signed-off-by: Dipak Shetty <dipak.shetty@zeiss.com>